Pass scrollWheelZoom option to Leaflet map
This commit adds the ability to set `data-scrollWheelZoom` on the map
DOM element and have that value passed through to the Leaflet map
object.

http://leafletjs.com/reference.html#map-scrollwheelzoom

Disabling scroll wheel zooming protects the user from inadvertently
zooming the map, and changing the search filter when they were
attempting to scroll the page to show additional search results. This is
important in a catalog index layout that positions the map above or
below the search result thumbnails.